update PodTemplate build
